Working on SelectMintsFromList and supporting functions, re-order zerocoin Denomination list with larger denominations first so it can be used for this function

+++ src/rpcwallet.cpp
@@ -2464,7 +2464,7 @@ Value spendzerocoin(const Array& params, bool fHelp)
 
     Object ret;
     ret.push_back(Pair("txid", wtx.GetHash().ToString()));
-    ret.push_back(Pair("bytes", wtx.GetSerializeSize(SER_NETWORK, CTransaction::CURRENT_VERSION)));
+    ret.push_back(Pair("bytes", (int64_t)wtx.GetSerializeSize(SER_NETWORK, CTransaction::CURRENT_VERSION)));
     ret.push_back(Pair("duration_millis", (GetTimeMillis() - nTimeStart)));
     Array arrSpends;
     for (CZerocoinSpend spend : vSpends) {
